# packages/nest-monitor/public/swagger.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/error/parse/{id}":{"get":{"operationId":"ErrorController_paseError","parameters":[{"name":"id","required":true,"in":"path","description":"异常ID","schema":{}}],"responses":{"200":{"description":""}},"tags":["error"]}},"/api/error/customParse":{"post":{"operationId":"ErrorController_customParseError","summary":"","description":"自定义异常解析","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/customParseDto"}}}},"responses":{"201":{"description":""}},"tags":["error"]}},"/api/error/add":{"get":{"operationId":"ErrorController_create","parameters":[],"responses":{"200":{"description":""}},"tags":["error"]}},"/api/error/analysis":{"get":{"operationId":"ErrorController_getErrorByType","parameters":[{"name":"type","required":false,"in":"query","description":"异常type 默认查所有","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","description":"结束时间(从当日23:59:59点开始)","schema":{"type":"string"}},{"name":"startDate","required":false,"in":"query","description":"开始时间(从当日00:00:00开始)","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["error"]}},"/api/error/{id}":{"get":{"operationId":"ErrorController_getErrorById","parameters":[],"responses":{"200":{"description":""}},"tags":["error"]}},"/api/error":{"get":{"operationId":"ErrorController_getErrors","parameters":[],"responses":{"200":{"description":""}},"tags":["error"]}},"/api/auth":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/user":{"post":{"operationId":"UserController_addUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"200":{"description":""}},"tags":["user"]},"get":{"operationId":"UserController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["user"]},"delete":{"operationId":"UserController_deleteUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"]}},"/api/project/auth/{key}":{"get":{"operationId":"ProjectController_getProjectIdByApiKey","parameters":[{"name":"key","required":true,"in":"path","schema":{}}],"responses":{"200":{"description":""}},"tags":["project"]}},"/api/project":{"post":{"operationId":"ProjectController_add","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectDto"}}}},"responses":{"201":{"description":""}},"tags":["project"]},"get":{"operationId":"ProjectController_getAll","parameters":[{"name":"access-token","in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["project"]},"delete":{"operationId":"ProjectController_deleteProById","parameters":[],"responses":{"200":{"description":""}},"tags":["project"]}},"/api/project/{id}":{"get":{"operationId":"ProjectController_getProById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["project"]}},"/api/sourcemap":{"post":{"operationId":"SourcemapController_uploadSourceMap","parameters":[{"name":"appid","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["sourcemap"]}},"/api/tag":{"post":{"operationId":"TagController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagDto"}}}},"responses":{"201":{"description":""}},"tags":["tag"]},"get":{"operationId":"TagController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["tag"]},"put":{"operationId":"TagController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTagDto"}}}},"responses":{"200":{"description":""}},"tags":["tag"]}},"/api/tag/{id}":{"delete":{"operationId":"TagController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["tag"]}},"/api/email/test":{"get":{"operationId":"EmailController_test","parameters":[],"responses":{"200":{"description":""}},"tags":["email"]}}},"info":{"title":"Nest Monitor","description":"前端异常监控","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"customParseDto":{"type":"object","properties":{"lineno":{"type":"number","description":"lineno"},"colno":{"type":"number","description":"colno"},"sourcemap":{"type":"string","description":"sourcemap"}},"required":["lineno","colno","sourcemap"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string","description":"username"},"password":{"type":"string","description":"password"}},"required":["username","password"]},"UserDto":{"type":"object","properties":{"username":{"type":"string","description":"username"},"password":{"type":"string","description":"password"},"email":{"type":"string","description":"email"}},"required":["username","password"]},"ProjectDto":{"type":"object","properties":{}},"CreateTagDto":{"type":"object","properties":{"name":{"type":"string","description":"name"},"value":{"type":"string","description":"value"}},"required":["name","value"]},"UpdateTagDto":{"type":"object","properties":{"id":{"type":"number","description":"id"},"name":{"type":"string","description":"name"},"value":{"type":"string","description":"value"}},"required":["id"]}}},"externalDocs":{"description":"swagger.json","url":"http://localhost:5003/swagger.json"}}